Global cleaning

+++ src/main/java/org/openstreetmap/osmaxil/dao/ElevationDataSource.java
@@ -7,19 +7,24 @@ import org.openstreetmap.osmaxil.Application;
 import org.openstreetmap.osmaxil.model.ElevationImport;
 
 public interface ElevationDataSource {
-	
-	public enum Type {DB, FILE};
-	
-	public enum Use {DTM, DSM};
-	
+
+	public enum Type {
+		DB, FILE
+	};
+
+	public enum Use {
+		DTM, DSM
+	};
+
 	static public final Logger LOGGER = Logger.getLogger(Application.class);
-	
+
 	abstract void init(String source, int srid);
-	
+
 	abstract public int getSrid();
-	
-	abstract public ElevationImport findElevationByCoordinates(double x, double y, float valueScale, int srid);
-	
-	abstract public List<ElevationImport> findAllElevationsByGeometry(String includingGeomAsWKT, String excludingGeomAsWKT, float valueScale, int shrinkRadius, int geomSrid);
+
+	abstract public ElevationImport findElevationByCoordinates(float x, float y, float valueScale, int srid);
+
+	abstract public List<ElevationImport> findAllElevationsByGeometry(String includingGeomAsWKT, String excludingGeomAsWKT, float valueScale,
+			int shrinkRadius, int geomSrid);
 
 }
