+++ src/main/java/org/openstreetmap/osmaxil/dao/ElevationRasterFile.java
@@ -23,7 +23,7 @@ public class ElevationRasterFile implements ElevationDataSource {
 	private double xPixelSize, yPixelSize;
 
 	private List<Band> bands;
-	
+
 	public ElevationRasterFile(String source, int srid) {
 		this.init(source, srid);
 	}
@@ -53,18 +53,18 @@ public class ElevationRasterFile implements ElevationDataSource {
 	}
 
 	@Override
-	public ElevationImport findElevationByCoordinates(double x, double y, float valueScale, int srid) {
+	public ElevationImport findElevationByCoordinates(float x, float y, float valueScale, int srid) {
 		int xFile = (int) Math.round((x - xUpperLeft) / this.xPixelSize);
 		int yFile = (int) Math.round((yUpperLeft - y) / this.yPixelSize);
-		double[] data = new double[1];
+		float[] data = new float[1];
 		bands.get(0).ReadRaster(xFile, yFile, 1, 1, data);
 		ElevationImport result = new ElevationImport(x, y, data[0] * valueScale);
 		return result;
 	}
 
 	@Override
-	public List<ElevationImport> findAllElevationsByGeometry(String includingGeomAsWKT, String excludingGeomAsWKT,
-			float valueScale, int shrinkRadius, int geomSrid) {
+	public List<ElevationImport> findAllElevationsByGeometry(String includingGeomAsWKT, String excludingGeomAsWKT, float valueScale, int shrinkRadius,
+			int geomSrid) {
 		List<ElevationImport> result = new ArrayList<>();
 		// TODO the task is not trivial:
 		// https://gis.stackexchange.com/questions/186483/how-to-get-the-pixels-from-a-geotiff-file-in-gdal-python-for-a-given-polygon
